backend/requirements.txt
# FastAPI 프레임워크
fastapi
uvicorn[standard]
mysql-connector-python
pydantic
aiomysql
databases

httpx
pytest
pytest-asyncio
asgi_lifespan
